refactor: move fiatRow into separate component
